Free dead Layouts and Types in Scavenger
Prior to this diff, we treated `Runtime::layouts_` as a normal root,
meaning we'd never free any `Layout`s or `Type`s (since every `Type` is
referenced by at least one `Layout`).

The high-level idea to fix this is to treat `layouts_` as weak roots, only
keeping a given `Layout` alive if it's referenced indirectly from an object's
`Header` or directly by a `Type`. There are a few extra considerations:
1. All builtin layouts are still treated as strong roots, since `Runtime` has
   functions to create instances of all of them.
2. `processLayouts()` runs last as part of a GC and is somewhat analogous to
   `processDelayedReferences()`: it updates forwarding references in `layouts_`
   where appropriate and replaces references to dead `Layout`s with `None`.

Based on Facebook D18821346