Avoid out-parameter usage in prepareCallableCall
Summary: Change `prepareCallableCall` and `prepareCallableDunderCall` to return two results in a struct instead of using 1 result and an `word*` out-parameter. This should make it easier to deal with the function in the assembly interpreter because the results are reported in 2 registers instead of a register and a memory location that must be allocated somewhere.

Based on Facebook D19134031