Export library files to build directory
This prepares pyro for distribution by consolidating the different places where we are pulling libraries from into a single directory in build. The libraries in the build directory will get automatically updates if they change in source. This will allow us to remove any use of the pyro source directory in the runtime and libraries and instead refer to the build/install path.

The copy is done by listing all the relevant files and directories from `pyro/library` and `pyro/third-party/cpython/Lib` and performing a `cp -r` into the build directory `build/Lib`. This makes sure that all the files as well as the directory modules (and their recursive contents) get copied into the build directory.

Based on Facebook D18717702