Shrink caches allocated per PC from 4 to 1
The previous diffs on this stack already prepares the interpreter to use only the first entry of 4 entries assigned to each PC.

Therefore, this diff allocates only 1 entry per PC, and adjusts all related constants with that to make the interpreter/assembly interpreter work with the change.

The impact of this change is to reduce the memory usage of caches 4 times. Also, data cache miss rate is expected to decrease, which should be contributing to CPU time reduction.

Based on Facebook D19042675