Avoid dropping unused C-API symbols
Currently, the linker drops any symbol that is not used in the python binary. Therefore, any external library trying to link against the binary might not find some C-API symbols.

This modifies the `extension-modules` library to not drop the symbols from the `extension` library. Furthermore, this had to remove the `runtime` dependency from `extension` as the `WHOLE_ARCHIVE` flag caused `runtime` to fail with multiple symbol errors. The only symbols that `extension` immediately required were from `mutex.h` so this pulls out the mutex sources into its own static lib that is added to `runtime` and `extension`

Finally, a couple of CPython files were failing to compile since they are still referring to symbols we don't support in our runtime (i.e. `PyList_Type`). Given that these are files of modules that we are not currently using, this just drops them from the list.

Based on Facebook D19143374