Avoid calling doCallFunction from other handlers
It's dangerous to call other opcode handles from a handler since now
we're rolling out opcode rewriting that assumes that a handler is
executing a specific opcode that handler is designed for. This change
introduces tailCallFunction to serve the purpose of calling
doCallFunction from other handlers.

Based on Facebook D19162305