Clear HandleScopes during Runtime destruction
Currently, `~Runtime()` is not able to free extension memory which is referenced by an active `HandleScope` on the stack. Make this more robust by clearing all `HandleScope`s before trying to free extension memory.

This is safe to do as long as we don't try to return from a stack frame which has active handles after destroying the runtime (currently this doesn't happen).

This solves the memory leak on D18821207 and should prevent similar leaks in the future.

Based on Facebook D19330383