Use PySys_SetArgv() in Py_BytesMain()
This sets the first entry in `sys.path` to be the path of the Python script being run.

This was previously blocked on pyro-perf not being a proper distribution.

Now that we have pyro-perf working, it's obvious that the managed implementation is too slow on the microbenchmarks (probably because it needs to import os.path). Revert to the native implementation developed in D18410142.

Based on Facebook D19333275