Add a few simple regression tests for Py_Main
As requested in D18898678.

I don't want to add too many tests here because they are fairly heavy and most of the logic should be tested by downstream tests. But this will prevent major features from regressing.

Note that this caught a bug in `Py_Main()` on mac machines - for now we just support ASCII arguments as this is good enough for the tests, and I don't expect anyone to actually use `Py_Main()` in practice (they should be using `Py_BytesMain()`

Based on Facebook D19333138