Implement str.expandtabs
Since nothing we know in Django yet calls `expandtabs`, I implemented it in managed code, using a `_strarray` to try to reduce allocations as much as possible.

If it ever shows up in our profile, we could very easily rewrite the implementation in native code.

Currently used in a few tests (including `test_ast`)

Based on Facebook D19277877