Implement bytes.__contains__ with byteslike argument
Summary: Many of CPython's tests call `bytes.__contains__` with a single `byte`. We can dispatch to `_byteslike_find_byteslike` for the general byteslike case. If `bytes.__contains__` ever shows up in a profile, we can add a specialized helper function that assumes `bytes` self and immediately returns a `bool` instead of an `int`.

Based on Facebook D19397953