Implement _Py_RestoreSignals
CPython's implementation hides the `SIGPIPE` and `SIGXFSZ` under `ifdefs`, however, seeing as how those signals are defined on all POSIX platforms, we should be able to safely set them without the `ifdefs`.

CPython also sets `SIGXFZ` under an `ifdef`, but our system doesn't support it, and it doesn't seem to exist in any platform, so this diff leaves it out.

Based on Facebook D19430982