Allow client reporting through diff number
Update the benchmark client to also report the benchmark results through the diff number instead of the phabricator version. This is needed for projects like cinder that live outside of fbcode and only have information about the current revision.

This also removes the deprecated NodeAPI uses and instead replace them with the latest Phabricator APIs.

Based on Facebook D19356484