Move sys init to SysModule class
The only part of initializing `sys` that must be done in the runtime is initializing instances (`sys.stderr`, `sys.stdout`, and `sys.displayhook`). Factor that out into its own function.

Moving `sys` initialization out of the runtime makes the module less of a special case, so we can declare it in `kBuiltinModules`.

Based on Facebook D19499880