Remove uses of reserveLayoutId as a builtin base in tests
Using `reserveLayoutId` for a value that needs to be a builtin layout ID is a
tricky business. The builtin base storage on `Type` assumes that builtin base
layout IDs are <256 and stores them in a byte. Now that we may have >255
classes created on boot, a test harness may reserve a bigger layout ID than we
have storage for.

This diff replaces reserveLayoutId with a dummy built-in layout ID.

Based on Facebook D19588666