Stop creating a method for every symbol
Summary: I am about to introduce a new scheme for identifiers. We will not be able to keep maintaining the current practice of creating one method for each symbol in the `Symbols` class. Luckily this feature isn't used too much (compared to how many symbols we have). So this change removes it.

Based on Facebook D19643477