Introduce ID(x) macro to name symbols
This renames the symbols from something like `SymbolId::kDunderAdd` to `SymbolId::k___add__`. It also introduces an `ID(x)` macro that allows to specify symbols by their natural form as they occur in python like `ID(__add__)`; the macro is also necessary to avoid `clang-tidy` complains.

This makes the code more compact, makes it easier to search for python identifiers in the code, should end all discussion on how to translate a python identifier to a C++ symbol name and will make it easier for scripts to create code involving symbols.

P125664675 is the script used to perform most of the rewrite and may be helpful when rebasing older changes.

Based on Facebook D19643551