Avoid using symbols for non-identifiers
I am about to introduce a new scheme to use identifiers, that will no longer work for things that are not identifiers (i.e. have spaces in the string or similar). This patch changes the couple symbols we had that were non-identifiers to use a `SmallStr::fromCStr` or allocate the strings on demand, as none of the cases is performance relevant.

Changed things to `Runtime::internStrFromCStr()` over `newStrFromCStr()` because it is closer to the symbol behavior and should avoid accidentally creating a lot of copies of the same simple string.

Based on Facebook D19643416