Add FUNC() macro to create C++ identifiers for builtin functions
Add `FUNC(modulename, name)` macro to create the C++ identifier to use for a native function in a given python module. i.e. to implement `builtins.__import__` in python as a builtin function you would use `FUNC(builtins, __import__)` to get a matching identifier.

This makes it easier to search for identifiers in the code; avoids discussions on how to translate python names into C++ names; and makes it easier to generate code.

Most of the conversion was performed by the script in P125729355.

Based on Facebook D19644154