Add METH(type, name) macro to name builtin methods
Add `METH(type, name)` macro to create the C++ identifier to use for a native method on a type. i.e. to implement `int.__hash__` in python as a builtin function you would use `METH(int, __hash__)` to get a matching C++ identifier.

This makes it easier to search for python identifiers; avoids discussions on how to translate python names into C++ names the simplicity of the scheme makes it easier to generate code.

Most of the conversion was performed by the script in P125723491.

Based on Facebook D19644369