Remove dictAtIfAbsentPut, dictValueCellAtPut
dictAtIfAbsentPut was only used by `dictAtPutInValueCellXXX` and I don't see any other upcoming users. This does not warrant the complicated callback mechanism. Instead implement the algorithm directly in `dictAtPutInValueCellByStr` and avoid the virtual function call and the extra callback types.

While on it, also remove the unused `dictAtPutInValueCell` and only leave `dictAtPutInValueCellByStr` around. Because this functionality is only used in the context of types and modules, we should never need the former variant for non-string keys.

Based on Facebook D19698725