Patch in internal Incremental{De|En}coder classes
Most of our `_io` classes call `getincrementaldecoder` or `getincrementalencoder`, which go through the `codecs` infrastructure to receive back an `IncrementalDecoder` or `IncrementalEncoder`. However, these classes just wrap around basic functionality that is already defined in `_codecs.py`.

This diff adds a copy of those classes in `_codecs` and adds a fast-path to them in `getincremental{en|de}coder`, to allow users to skip the `encodings` architecture, and instantiate `_io` classes early on in the runtime's initialization.

Once we can freeze directories into our runtime and can import `encodings` earlier on, we can remove these patches and use the `codecs` classes directly.

Based on Facebook D19563148