Use new builtins system
Use the new system from D18039226 to define all builtin functions and methods.

The new system will no longer pre-populate the module namespaces with (half-initialized) function objects for the builtins. We have to tweak some things for this:
- Have cacheBuildClass and cacheDunderImport insert a None value because the name does not exist yet when they run. This is not a problem as the ValueCell connected to the name will updated as usual as soon as the real functions are created.
- We have to wait in `builtins-module.cpp` until after the bytecode ran before we can assign intrinsic ids.

Most of this change was performed by the script in P125847156.

Based on Facebook D19699012