Remove using-directives
This is both error prone and forbidden by the style guide.

https://google.github.io/styleguide/cppguide.html#Namespaces

This change moves unit tests that were in the the `py` namespace
or a namespace nested under `py` into a innermost namespace
called `testing`.

In one case, this revealed a test helper defined in
`ic-test.cpp`, a test file, being declared in `ic.h`, an
implementation header file.

Finally, the file with the `main` function for running unit tests
now explicitly imports the symbols it needs from the googletest
namespace into the anonymous namespace declared in the file.

Based on Facebook D19756031