Add HeapProfiler
Our runtime lacks introspection for where memory is allocated, how much memory
is allocated, and what is stored there. In order to gather this data, this
stack of diffs piecemeal ports the Dart heap profiler ([dart]), Hotspot heap profiler
([hotspot]), and ART heap profiler ([art]) to Pyro.

[dart]: https://github.com/dart-lang/sdk/commit/dcaf11db9649fa4a6af7f97d4c3c8d512a2c91ee
[hotspot]: https://github.com/sourcemirror/jdk-9-hotspot/blob/master/src/share/vm/services/heapDumper.cpp
[art]: https://android.googlesource.com/platform/art/+/master/runtime/hprof/hprof.cc

This heap profiler scans the heap and writes out objects in the HPROF file
format, which is a widely used file format in the Java world. Many tools such
as JHAT and MAT accept input in HPROF and can be used to perform analysis on
Pyro's heap.

When it is complete, the profiler should be flexible enough to easily be used:

* from user code (by calling a `heap_dump(...)` function)
* internally in the runtime if it hits some weird [edge case] (like a coredump)
* if the runtime receives some kind of signal like SIGUSR1

[edge case]: https://github.com/dart-lang/sdk/commit/a7675e8dd3ea172e7fa7c471ea151a38e4f82bfc

and many more to-be-explored cases.

Based on Facebook D19731259