Add array.array.__new__ and array.typecode
Implements a basic version of `array.array.__new__` that doesn't try to add the `initializer` to the newly constructed array. Further functionality will be added in further diffs.

Since we know that distillery doesn't make much use of `array.array`, we can leave most of the initialization code in managed code, and rewrite parts as we see fit if they ever come up in profiles.

This also adds the `array.typecode` property for easier testing, and removes the unnecessary `array.typecodes` method (since it shouldn't be a method on `array.array`)

Based on Facebook D19663469