Add _sre module
Adds the `_sre` module to our runtime. `clang-tidy` is unhappy with the `sre*.h` headers which don't `include` anything and don't compile by themselves, so this diff just copies over `_sre.cpp` into `/ext/Modules/_sre`, and inlines all the headers. We can get away with this since our end goal is to rewrite a lot of `_sre.cpp` anyway.

This diff also removes references to the `USE_BUILTIN_COPY` macro since it was an experimental opt-in feature in 3.6 that got removed in 3.7+.

This allows us to add a patch in an upcoming diff to handle non-ASCII characters and unblock distillery.

Based on Facebook D19728331