Remove xxx_ClearFreeList() implementations
The API is not part of the PEP384 API and as far as I can tell the only user of it is the cpython gc module. We will not use the cpython gc module, but instead create our own so we can just as well remove the API.

Removed two test files that happened to no longer contain any tests after this change.

Based on Facebook D19882888