Patch _sre to allow non-ascii
In order to unblock distillery, this diff adds a quick patch to `state_init` that decodes all non-ASCII strings to be `UCS4` strings. This allows us to use all of CPython's `_sre` infrastructure with non-ASCII strings without needing to rewrite anything to be UTF-8 aware.

It also removes the call to `PyUnicode_KIND` in `getstring` (since `PyUnicode_KIND` is the current smoke signal in our interpreter that something is trying to access a string's internals).

Based on Facebook D19856965