Add support for passing arguments to the interpreter in the benchmark server
Add an optional argument, '-a/--interpreter-args', to supply command-line
arguments that should be passed to the python interpreter when running
a benchmark. The arguments are passed as a single string (e.g.
`-a "-X jit -X jit-no-frame"`). Multiple values may be supplied; they apply
to each interpreter, in order.

Based on Facebook D20047922