Terminate runtime for unsupported cache invalidation on updating existing type attr
Currently the Pyro runtime terminates itself when `__setattr__` or
`__getattribute__` get updated on a type even when they do not exist in
the type. With this change the runtime gets terminated only when they
are populated in the updated type since in case of the absence of
those attributes, they cannot be cached and no cache invalidation is
needed from the first place.

Based on Facebook D20037333