Check that runtime constants match C-API
We define some constants in both the runtime and the C-API. We try to avoid including the C-API headers from the runtime or the other way round so some duplication is unavoidable. This adds some `static_assert`s to ensure the duplicated values stay in sync.

This also moves some constants from version.cpp to version.h so they become compiletime constants.

Based on Facebook D20002606