Rework emulation of PyCFunction
This changes how `PyCFunction_NewEx` and `PyModule_Create2` create
functions.

Background
==========
CPython uses a special `PyCFunction` object in these situations. A
`PyCFunction` is a pair of a `self` value and a pointer to a C function,
where `self` is a pointer to another python object or `NULL`. Contrary
to function objects, a `PyCFunction` does not provide a `__get__` method,
and therefore should not create a `BoundMethod` when querying from a
class instance.

We do not want to introduce another object type to represent
`PyCFunction`, hoping to keep things performant without having to
introduce additional special cases in the interpreter. Instead, we employ
special function trampolines.

`PyCFunction_NewEx` used a `moduleTrampolineXXX` that would make sure
the `self` value is passed on to the C/C++ function. Strangely, the
`self` value was not stored as part of the function object, and instead
it was just assumed to always be equal to the module containing the
object (which was mostly but not necessarily always true).  Using a
function object would also show different binding behavior when
`__get__` was called, which I assume motivated a 2nd code path in the
function using `methodTrampolineXXX` that would not automatically pass
in a self parameter contained in the function, instead,
`PyCFunction_NewEx` would construct a BoundMethod and bind the self
parameter this way.

Changes
=======

We can simplify and streamline things by always using a bound method
with a `methodTrampoline`. As this can bind a self value and will result
in the desired descriptor binding behavior.

- Introduce `newCFunction` helper that creates a function with a
  `methodTrampolineXXX` and always wraps the self parameter with a
  BoundMethod.
- Translate a nullptr self to `_Unbound` in the `BoundMethod` and back
  to nullptr in the trampoline.
- This also fixes the code wrongly assuming the `module` parameter is a
  module object, while it actually is a module name that is supposed to
  be used to fill in the `__module__` attribute of the result.
- Use this simpler and streamlined `newCFunction` for both
  `PyCFunction_NewEx` and `PyModule_Create2`.
- Remove the module trampoline.

This just creates a `PyCFunction` object which basically is a
pair of a `self` value and a function pointer to a C function. Contrary
to `PyFunction` objects it does not feature a `__get__` method so the
binding behavior is slightly different.

Based on Facebook D20119536