Add encodings files to pyro
Since we can't bind builtin functions as static methods, we need to fork over encodings files for each Codec we look up. This adds the UTF-16 encodings as well as a few unicode_escape ones. While we don't have all the encoders/decoders needed to instantiate the `Codec` objects, this means that calls to `codecs.getencoder()`, will fail with the missing encoder/decoder instead of a more difficult to figure out error.

This adds the files exactly as they are from CPython

Based on Facebook D20197339