Move module state from C-API handle table to heap object
This change stores the C-allocated module state in Pyro-managed module objects to restore it after the module's entry is gone from the C-API handle table.

Note that the module state is deallocated only for module objects that show up either in `Runtime::modules()` or `Runtime::modulesByIndex()`. We assume that modules with state are registered to the interpreter state by `PyState_AddModule`, which is really the case for all our extension modules.

Based on Facebook D20052361