Generate Unicode database for str.islower
This diff pulls in just enough of `makeunicodedata.py` to generate the type
records needed for `str.islower`. Implementation of `_PyUnicode_IsLowercase`
has been left for when we can fully eliminate `unicodectype.c`.

We modify the script both to take advantage of Python 3.7 features and to
generate a cleaner interface for Pyro. We add `unicode-db.h` for types and
function declarations, and we add `unicode-db.cpp` for static arrays used by
those functions. We maintain `unicode.h` and add `unicode.cpp` for a simpler
interface, adding a level of abstraction between the database and our runtime.

Leave the name and property databases for later, as they are not yet necessary.

Facebook:
@allow-large-files

Based on Facebook D20121336