Add workaround so &PyXXX_Type still works with gcc in C mode
We do some trickery for `PyXXX_Type` in pyro and define them as a macro, example for PyLong_Type:
```
#define PyLong_Type  *(PyLong_Type_Func())
PyTypeObject* PyLong_Type_Func();
```
Most users use this as `&PyLong_Type_Func` which results in a `&*(PyLong_Type_Func())` expression. This should work fine even though `PyTypeObject` itself is an incomplete type. However it seems gcc has a bug rejecting this pattern in C code (interestingly not in C++ mode though).

See also: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=88827 which will be fixed in the upcoming gcc10.

As a workaround this adds a definition for GCC in C mode so it no longer rejects the types in C mode.

Based on Facebook D20265381