Replace _multiprocessing stub with C file
Now that we've PEP-384'd `_multiprocessing`, we could use its full functionality instead of relying on the stub.

In CPython, `semaphore.c` does not need to `include <sys/time.h>` to run `gettimeofday`, presmuably because some other header file already included it. However, in our build script, we don't link the library when trying to build `semaphore.c`. Therefore, this diff also adds an `include` at the top of the file.

Based on Facebook D20231958