Return None if a type's member string is NULL
`_multiprocessing` causes a segfault when trying to access the `name` field of a `SemLock`. This is caused by the fact that it's `name` field should be `NULL` when `unlink` is `True`. Unfortunately, our implementation of `_get_member_string` tries calling `newStrFromCStr` on the field, which causes the crash.

We can fix this by more closely mimicking CPython's `PyMember_GetOne` implementation and returning `None` if the address points to `NULL`.

Based on Facebook D20392303