Switch library to use CPython's namedtuple
This change switches the library to use the CPython version of
`namedtuple` instead of the custom version which was based on
`PyStructSequence`, reducing the divergence of our standard library.

This also makes `namedtuple` more compatible by removing some user-visible
attributes on the types needed by `_structseq_field`.  It also adds
user-visible attributes on the type that should have been present.

To make it easier to bootstrap `namedtuple`, we have used f-strings,
which are standalone, in place of `string.format`, which has many
runtime dependencies.

This change replaces Pyro-internal implementation of `_namedtuple` for CPython's one.

**Side note**

Huge thanks to @matthiasb for implementing `string.format` and the
inspiration for this change to sync our `namedtuple` with CPython's.

Based on Facebook D20263188