Use merge sort for list.sort
This change uses merge sort for `list.sort` to enable it to sort symbols for compiling a big python file (e.g., ttypes.py generated for Thrift).

This code can be further optimized not to call `__lt__` for a set of common builtin types (e.g., `str`).

Based on Facebook D20348106