Implement str.translate(str)
This implements `str.translate(str)` for `numpy`.

It uses a fastpath when passed an ASCII only string, since we know that that means that each byte in the source will correspond to exactly one byte in the result, otherwise it goes through a `__getitem__` lookup in a `try ... except` in managed code.

Based on Facebook D20430079