Implement object.__init_subclass__
This method is always a `classmethod`, and it is implicitly made so in `typeInit()`.
We have to be explicit about `object.__init_subclass__` because the `object` type
uses `bootstrap=True`, which currently does not call `typeInit()`.

Calling `super().__init_subclass__()` requires a `_super()` helper for
bootstrapping type before the `super()` class has been created.

Based on Facebook D20490910