Add more built-in functions to implement methods on module_proxy
The motivation of this change is to avoid using `_Unbound` as a
sentinel value within the methods of `module_proxy`.  This allows
us to return the value of a key when the value is `_Unbound`.

As part of this change, the `_module_proxy_guard` method has been
replaced by an implicit guards in the built-in functions to avoid an
extra method call.

Based on Facebook D20549196