Port D15454859 to third-party/cpython
D15454859 introduced a new `_myreadline` module object and moved `_PyOS_ReadlineTState` into the module state as we usually do when preparing code for PEP384.

This unfortunately ignored the fact that `pythonrun.h` still declared it as a global variable and `Modules/readline.c` was using it. Causing the optional readline module to fail compilation. It was not recognized at the time because optional modules not compiling is a common occurences and won't mark the CI tests as failed.

Upon inspection I decided to revert parts of D15454859 ("[Port 3.8 to 3.6] Cleanup accessors of PyThreadState"):

The problem here is that unfortunately the readline library itself has global state. `_PyOS_ReadlineLock` and `_PyOS_ReadlineTState` were designed to protect that global state and synchronize users of the the single global callback we can register with the library.

This is a rare case where we should not have done our usual PEP384 changes that move everything into the module state, because the global variables were protecting and synchronizing access to the global readline state.

Original Diff: D20607069

Based on Facebook D20607393