Add Unicode name db
Add the Unicode name data needed to get a code point from its name and vice versa.

As in CPython, we split names up into their component words,
sorting the most common by descending frequency and the rest by length.
We store the offsets of each word, and map to these offsets in a phrasebook.

Hangul syllables and CJK ideographs are special-cased.

Based on Facebook D20262850