Use CMAKE_TOOLCHAIN_FILE
We should be using `-DCMAKE_TOOLCHAIN_FILE=util/xxx.cmake` rather than `-C util/xxx.cmake`.

This has the advantage that the platform/toolchain file is re-executed with every cmake run so we don't get stale value in the cache (for the stuff in those files at least) and things added between revisions will be picked up automatically without the need to clean the cmake cache.

This changes the files to be compatible with `CMAKE_TOOLCHAIN_FILE` and adds a message to the main cmake file to help users that did not specify a `CMAKE_TOOLCHAIN_FILE`.

Based on Facebook D20648529