Improve strSlice to linear time
The existing implementation of `Runtime::strSlice()` is quadratic because it
loops through the whole string to find the correct offset for each character.
We can make this linear by finding offset relative to the current code point
rather than the beginning of the string every time.

Move `numTrailBytes()` to a new `UTF8` class, and make it return the full
length of the code point rather than just the number of trailing bytes.

Based on Facebook D20776651