Port 3.7 changes to the regex module
This diff ports over all the necessary changes in `_sre.cpp` by applying the changes to each `*re*` file in `Modules/` to the apporpriate inlined section, and making sure that the result is still PEP-384 compliant.

Since the `re` module is a Python library however, this diff also ports over the changes in each of the `*re*` files in `Lib/` This is necessary because `re` checks a magic number against the `_sre` magic number, and we can't use `re` without the magic numbers corresponding.

Since we need CPython to build and run as well however, we also need to port over the `Modules/` changes to go along with the `Lib/` changes.

The new `re` module now makes us go over the cache rewriting limit, which prints an error to `stderr`. This diff also modifies a few tests in `UnderWarningsExtensionApiTest` which capture `stderr` to make sure that they only check the ending of the stream (to exclude the `cache rewriting limit hit` warning).

Based on Facebook D20799541