Implement object.__class__ setter
This adds a setter for the `__class__` attribute on instances. To accomplish
this, it pulls out the layout of the instance, creates a copy the the layout,
and sets the described type of the copy to the new type. This layout is also
registered in a new Runtime table of cached transitions between layouts.

This diff also adds checks on the types to make sure that we can do a safe
`__class__` assignment.

This diverged from CPython since it looks at `Py_TPFLAGS_HEAPTYPE` to determine
if it can set `__class__`. In place of it, this disallows any assignments to a
builtin type (since all CPython builtin types are static).

Based on Facebook D19396302