Add Hangul predicates
Implement predicates to check if a code point is a Hangul syllable or jamo.
Conform to the Unicode terminology of 'lead', 'vowel', and 'trail' jamo classes.

Follow comments on D20843451 for structure of predicates.

Move Unicode constants into the `Unicode` class,
reversing the dependency between `unicode.h` and `unicode-db.h`.

Based on Facebook D20871361