Remove readline stub module
This removes `library/readline.py` which only contained stubbed out functions. It does not appear to be used at the moment.

This improves the experience with `site.py` and `doctest.py` which otherwise attempt to call the stubbed out functions but do not try to do so if the readline module fails to import.

Based on Facebook D20955032