Set filesystemencodeerrors to surrogatepass
The current implementation of `_os.open` just uses the bytes of our string representation of the filenames as-is. While some C-API code encodes the filename with `Py_FileSystemDefaultEncoding` / `Py_FileSystemDefaultEncodingErrors` (aka `sys.getfilesystemdefaultencoding`, `sys.getfilesystemdefaultencodingerror`).

This switches the default encoding errors to `surrogatepass` so the `_os.open` results match up with cpython (from the posix module) when there are surrogates in the filename.

Based on Facebook D20957902