Do not fail in doctest without sys.gettrace/sys.settrace
Patch doctest to not install a trace handler if `sys.gettrace`/`sys.settrace` do not exist. The trace function was only used to detect entry/exit from a debugger and revert changes to `sys.stdxxx` while doing so, but we can easily live without that for now.

This also adds an `encoding` member to our `sys.stdxxx` files with the value `"utf-8"`.

Based on Facebook D20955128