Store atexit function and void* context on the runtime instead of PyObject*
This diff changes `_Py_PyAtExit` to create a thunk (function + module pair)
that is stored on the runtime with an opaque pointer. The runtime then calls a
pointer to a function in `pylifecycle.cpp` with the thunk, which can unpack it.
This has a drawback of allocating but insulates the runtime about
lifecycle/atexit implementation details.

The runtime gained a `PyObject*` as part of D20774186.

Based on Facebook D20996870