Implement ctypes._SimpleCData constructor
CPython makes use of the `stgdict` throughout `ctypes`. This is a data structure that pyro doesn't plan on supporting if we don't need to. Because of this, a few changes have been made to the initializer:

In CPython, the `_SimpleCData.__new__` constructor makes sure that the passed in `cls` has a `stgdict`. Concretely, this is a check to make sure that the user isn't trying to initialize a `_SimpleCData`, and is instead initializing a subclass of it (which would have been created using the `PyCSimpleType` metaclass). Our equivalent for that would be to make sure that `cls` is strictly a subclass of `_SimpleCData` which has a `_type_` field.

It then `malloc`s enough memory to fit the value needed by the class to point to. However, since our `ctypes` implementation is all in managed code, we can just store any data needed in a hidden field on the class and store our own pyro objects in that field.

In CPython, the `_SimpleCData.__init__` constructor tries getting the `stgdict` out and calling its `setfunc` method on the given value. The `setfunc` method however is standardized based upon the format (found in `_type_`), which allows us to do the appropriate error checking/casts based off of a switch on the `_type_`.

Finally, CPython stores a reference on each new object of the object used to create it. This is a manual reference count attempt to keep track of what memory still has objects pointing to it. Since we have an automatic garbage collector, we can ignore the manual reference count need and simply trust that our garbage collected will free the memory when it needs to.

Based on Facebook D20923554