Implement memoryview.obj
Summary: `memoryview`s in CPython store the original object that they were created with inside of an attribute called `obj`. This means we need to add a new field to `memoryview`s to store the original object without needlessly complicating the logic for getting/setting items.

Based on Facebook D20999850