Implement _CDataType.from_buffer
Similarly to our decision in D20923554, pyro will rely on our managed types' `__dict__` and our own garbage collector instead of re-implementing CPython's functionality.

`_CDataType.from_buffer` should store the `Pointer` object in the `_value` field of a class instantiation, along with the offset that the getter shuld read memory from, and the getter should be able to interpret that pointer as the data for the type it should store (determined by `_type_`). Since django seems to only use `uint16` for now, this diff only implements that case.

Based on Facebook D20925694