Fix tuple overflow bug in layoutSetDescribedType
There was a bug in calculating the necessary length for the transition table; we only took into account the starting index, not the need to place three items.

The transition table starts at 3 items, but will get resized to 16, which is not a multiple of 3, so this code breaks.

Based on Facebook D21066474