Move some calling tests from moduleobject-test.cpp to methodobject-test.cpp
`PyModule_Create` is using the same underlying code to create functions as `PyCFunction_NewEx`. So it seems appropriate to test all the details (different `METH_XXX` flags) in the tests for `PyCFunction_NewEx` and not test them again in the context of `PyModule`.

This also removes some tests: We make sure every `METH_XXX` combination is tested once to ensure it works. But we no longer test every error path and call type since we already do that in `trampolines-test.cpp` for the `methodXXX` trampolines.

`MethodWithClassFlagRaisesException` is not moved as that behavior is specific to `PyModule_Create`.

Based on Facebook D21074859