Do not hardcode layout id values in test
Dumping the flags field in the `DumpExtendedInstanceWithOverflowDict` included dumping the `builtin_base` layout-id that was encoded in that field. This means the test would need an update every time layout ids change and it is broken right now.

Change the test to no longer hardcode the result.

Based on Facebook D21133108