Seal instance layouts for type and module (and their subtypes)
Add a flag to `RawType` which, when set, will cause a type and all of its subtypes to have sealed layouts. Apply this flag to `type` and `module`

Right now, instance **layouts** for `type`, `module`, and their subclasses (such as metaclasses) are set up with an empty tuple in the overflow attributes, but type **instances** have no overflow tuple. This is confusing and also causes issues with our current debugging tooling - Right now, if you `po` any class with a custom metaclass, the debugger will abort.

Based on Facebook D21211729