Implement COMPARE_IN_* caching opcodes
This change implements a group of opcodes to cache & specialize `COMPARE_OP` with `CompareOp::IN` operation.

Note that this change does not include opcodes for `CompareOp::NOT_IN` operation and delegates its work to the future compiler opcode expansion to compile `COMPARE_OP` into `COMPARE_IN` and `NOT` operation.

Based on Facebook D19603005